roll back abandonment of basic_test_func
