+++ tests/test_isobased.py
@@ -6,7 +6,6 @@ from astropy import units as u
 from geovectorslib import geod
 
 import tests.basic_test_func as basic_test_func
-from tests.test_direct_power_method import TestDPM
 import WeatherRoutingTool.utils.formatting as form
 from WeatherRoutingTool.constraints.constraints import LandCrossing, WaveHeight
 from WeatherRoutingTool.ship.shipparams import ShipParams
@@ -196,7 +195,7 @@ def test_get_delta_variables_last_step():
 
     ##
     # initialise boat
-    tk = TestDPM.create_dummy_Direct_Power_Ship("simpleship")
+    tk = basic_test_func.create_dummy_Direct_Power_Ship("simpleship")
     tk.speed = boat_speed
     tk.use_depth_data = False
