+++ tests/test_route_postprocessing.py
@@ -7,7 +7,7 @@ import sqlalchemy as db
 from astropy import units as u
 from shapely.geometry import box, LineString, Point
 
-from tests.test_direct_power_method import TestDPM
+import tests.basic_test_func as basic_test_func
 from WeatherRoutingTool.constraints.route_postprocessing import RoutePostprocessing
 from WeatherRoutingTool.routeparams import RouteParams
 from WeatherRoutingTool.ship.shipparams import ShipParams
@@ -126,7 +126,7 @@ class TestRoutePostprocessing:
             starttime_per_step=start_time,
             ship_params_per_step=sp
         )
-        boat = TestDPM.create_dummy_Direct_Power_Ship("simpleship")
+        boat = basic_test_func.create_dummy_Direct_Power_Ship("simpleship")
         boat.set_boat_speed(6)
         with engine.connect() as conn:
             postprocessed_route = RoutePostprocessing(rp, boat, db_engine=conn.connection)
